IconChangeLinksController, OverlayTextsControllerをAdminネームスペースで囲った Adminが操作するコントローラなため。 依存関係を明示するため。
